MINE Type을 관리하는 ContentType enum을 http 패키지로 이동
