+++ src/main/java/codesquad/ClientHandler.java
@@ -1,5 +1,6 @@
 package codesquad;
 
+import codesquad.http.ContentType;
 import codesquad.http.HttpRequest;
 import codesquad.http.HttpResponse;
 import org.slf4j.Logger;
@@ -102,37 +103,5 @@ class ClientHandler implements Runnable {
 
     private static String getMimeType(String filePath) {
         return ContentType.getMimeType(filePath);
-    }
-
-    public enum ContentType {
-        HTML("html", "text/html"),
-        CSS("css", "text/css"),
-        JS("js", "application/javascript"),
-        ICO("ico", "image/x-icon"),
-        PNG("png", "image/png"),
-        JPEG("jpeg", "image/jpeg"),
-        SVG("svg", "image/svg+xml");
-
-        private final String extension;
-        private final String mimeType;
-
-        ContentType(String extension, String mimeType) {
-            this.extension = extension;
-            this.mimeType = mimeType;
-        }
-
-        public static String getMimeType(String filePath) {
-            String extension = getExtension(filePath);
-            return Arrays.stream(values())
-                    .filter(ct -> ct.extension.equalsIgnoreCase(extension))
-                    .map(ct -> ct.mimeType)
-                    .findFirst()
-                    .orElse("text/plain");
-        }
-
-        private static String getExtension(String filePath) {
-            int dotIndex = filePath.lastIndexOf('.');
-            return (dotIndex > 0) ? filePath.substring(dotIndex + 1) : "";
-        }
     }
 }
